+++ flexmeasures/data/schemas/scheduling/__init__.py
@@ -1,4 +1,4 @@
-from marshmallow import Schema, fields
+from marshmallow import Schema, fields, validate
 
 from flexmeasures.data.schemas.sensors import SensorIdField
 from flexmeasures.data.schemas.units import QuantityField
@@ -10,7 +10,22 @@ class FlexContextSchema(Schema):
     """
 
     ems_power_capacity_in_mw = QuantityField(
-        "MW", required=False, data_key="site-power-capacity"
+        "MW",
+        required=False,
+        data_key="site-power-capacity",
+        validate=validate.Range(min=0),
+    )
+    ems_production_capacity_in_mw = QuantityField(
+        "MW",
+        required=False,
+        data_key="site-production-capacity",
+        validate=validate.Range(min=0),
+    )
+    ems_consumption_capacity_in_mw = QuantityField(
+        "MW",
+        required=False,
+        data_key="site-consumption-capacity",
+        validate=validate.Range(min=0),
     )
     consumption_price_sensor = SensorIdField(data_key="consumption-price-sensor")
     production_price_sensor = SensorIdField(data_key="production-price-sensor")